clean up where tallying should be done
