+++ bluemira/codes/openmc/make_csg.py
@@ -10,8 +10,7 @@ in this module are in SI (distrance:[m]) unless otherwise specified by the docst
 
 from __future__ import annotations
 
-from collections.abc import Callable
-from dataclasses import dataclass, fields
+from dataclasses import dataclass
 from itertools import chain, pairwise
 from typing import TYPE_CHECKING
 
@@ -66,6 +65,196 @@ if TYPE_CHECKING:
 
 # Found to work by trial and error. I'm sorry.
 SHRINK_DISTANCE = 0.0005  # [m] = 0.05cm = 0.5 mm
+
+
+@dataclass
+class CellStage:
+    """Stage of making cells."""
+
+    blanket: list[openmc.Cell]
+    divertor: list[openmc.Cell]
+    tf_coils: list[openmc.Cell]
+    cs_coil: openmc.Cell
+    plasma: openmc.Cell
+    radiation_shield: openmc.Cell
+    ext_void: openmc.Cell
+    universe: openmc.region.Intersection
+
+    bounding_box: tuple[float, float, float, float]
+    half_bounding_box: tuple[float, float, float, float]
+    blanket_cell_array: BlanketCellArray | None = None  # only for default case
+    divertor_cell_array: DivertorCellArray | None = None  # only for default case
+
+    @property
+    def cells(self):
+        """Get the list of all cells."""
+        return (
+            *chain.from_iterable((*self.blanket, *self.divertor)),
+            *self.tf_coils,
+            self.cs_coil,
+            self.plasma,
+            self.radiation_shield,
+            self.ext_void,
+        )
+
+    @classmethod
+    def from_pre_cell_stage(
+        cls,
+        pre_cell_stage: PreCellStage,
+        tokamak_dimensions: TokamakDimensions,
+        material_library: MaterialsLibrary,
+        csg: BluemiraNeutronicsCSG,
+        *,
+        control_id: bool = False,
+    ) -> CellStage:
+        """
+        Make CellStage from PrecellStage
+
+        Parameters
+        ----------
+        pre_cell_stage:
+            Precell Stage containing blanket and divertor pre-cells
+        tokamak_dimensions:
+            A parameter
+            :class:`bluemira.radiation_transport.neutronics.params.TokamakDimensions`,
+            Specifying the dimensions of various layers in the blanket, divertor, and
+            central solenoid.
+        control_id: bool
+            Whether to set the blanket Cells and surface IDs by force or not.
+            With this set to True, it will be easier to understand where each cell came
+            from. However, it will lead to warnings and errors if a cell/surface is
+            generated to use a cell/surface ID that has already been used respectively.
+            Keep this as False if you're running openmc simulations multiple times in one
+            session.
+
+        Returns
+        -------
+        CellStage
+        """
+        # determine universe_box
+
+        z_max, z_min, r_max, r_min = pre_cell_stage.half_bounding_box()
+
+        z_min_adj = z_min - D_TOLERANCE
+        z_max_adj = z_max + D_TOLERANCE
+        r_max_adj = r_max + D_TOLERANCE
+
+        rad_shield_wall_tk = tokamak_dimensions.rad_shield.wall
+
+        # make the universe box, incorporates the radiation shield wall
+        universe = make_universe_box(
+            csg,
+            z_min_adj - rad_shield_wall_tk,
+            z_max_adj + rad_shield_wall_tk,
+            r_max_adj + rad_shield_wall_tk,
+            control_id=control_id,
+        )
+
+        blanket = BlanketCellArray.from_pre_cell_array(
+            pre_cell_stage.blanket,
+            material_library,
+            tokamak_dimensions,
+            csg,
+            control_id=control_id,
+        )
+
+        # change the cell and surface id register before making the divertor.
+        # (ids will only count up from here.)
+        if control_id:
+            round_up_next_openmc_ids()
+
+        divertor = DivertorCellArray.from_pre_cell_array(
+            pre_cell_stage.divertor,
+            material_library,
+            tokamak_dimensions.divertor,
+            csg=csg,
+            override_start_end_surfaces=(blanket[0].ccw_surface, blanket[-1].cw_surface),
+            # ID cannot be controlled at this point.
+        )
+
+        # make the plasma cell and the exterior void.
+        if control_id:
+            round_up_next_openmc_ids()
+
+        cs, tf = make_coils(
+            csg,
+            r_min - tokamak_dimensions.cs_coil.thickness,
+            tokamak_dimensions.cs_coil.thickness,
+            z_min_adj,
+            z_max_adj,
+            material_library,
+        )
+        # make the radiation shield wall
+        # which is a hollow of the universe box
+        rad_shield = make_radiation_shield_box(
+            csg,
+            z_min_adj,
+            z_max_adj,
+            r_max_adj,
+            universe,
+            material_library,
+        )
+        plasma, ext_void = make_void_cells(
+            csg,
+            universe=universe,
+            blanket=blanket,
+            divertor=divertor,
+            central_solenoid=cs,
+            tf_coils=tf,
+            rad_shield=rad_shield,
+            control_id=control_id,
+        )
+
+        cell_stage = cls(
+            blanket=blanket.get_hollow_merged_cells(),
+            divertor=divertor.get_hollow_merged_cells(),
+            tf_coils=tf,
+            cs_coil=cs,
+            plasma=plasma,
+            radiation_shield=rad_shield,
+            ext_void=ext_void,
+            universe=universe,
+            bounding_box=pre_cell_stage.bounding_box(),
+            half_bounding_box=pre_cell_stage.half_bounding_box(),
+            blanket_cell_array=blanket,
+            divertor_cell_array=divertor,
+        )
+        cell_stage.set_volumes(blanket, divertor)
+
+        return cell_stage
+
+    def set_volumes(
+        self, blanket_array: BlanketCellArray, divertor_array: DivertorCellArray
+    ):
+        """
+        Sets the volume of the voids. Not necessary/ used anywhere yet.
+        """
+        ext_vertices = exterior_vertices(self.blanket, self.divertor)
+        total_universe_volume = (
+            #  top - bottom
+            (self.universe[0].surface.z0 - self.universe[1].surface.z0)
+            * np.pi
+            * self.universe[2].surface.r ** 2  # cylinder
+        )  # cm^3
+
+        # is this needed?
+        # self.universe.volume = total_universe_volume
+
+        outer_boundary_volume = to_cm3(
+            polygon_revolve_signed_volume(ext_vertices[:, ::2].T)
+        )
+        ext_void_volume = total_universe_volume - outer_boundary_volume
+        if self.tf_coils:
+            for coil in self.tf_coils:
+                ext_void_volume -= coil.volume
+        if self.cs_coil:
+            ext_void_volume -= self.cs_coil.volume
+        self.ext_void.volume = ext_void_volume
+        blanket_volumes = sum(cell.volume for cell in chain.from_iterable(blanket_array))
+        divertor_volumes = sum(
+            cell.volume for cell in chain.from_iterable(divertor_array)
+        )
+        self.plasma.volume = outer_boundary_volume - blanket_volumes - divertor_volumes
 
 
 def is_monotonically_increasing(series):
@@ -2161,308 +2350,3 @@ class DivertorCellArray:
         return [
             openmc.Cell(region=stack.get_overall_region()) for stack in self.cell_array
         ]
-
-
-TALLY_FUNCTION_TYPE = Callable[
-    [list[openmc.Material], BlanketCellArray, DivertorCellArray],
-    tuple[
-        str,
-        str,
-        list[openmc.CellFilter | openmc.MaterialFilter | openmc.ParticleFilter],
-    ],
-]
-
-
-def filter_cells(
-    material_list,
-    blanket_cell_array: BlanketCellArray,
-    divertor_cell_array: DivertorCellArray,
-):
-    """
-    Create scores and the filter for the scores. Give them names.
-
-    Returns
-    -------
-    TBR
-        Achieved by (n,Xt) reaction, which counts the number of tritium-producing
-        nuclear reactions per neutron emitted at the source.
-
-        We used the (n,Xt) score because the Lithium produces a maximum of 1 Tritium per
-        reaction, so there won't be any concerns about uncer-counting the TBR.
-
-    Powers
-        Measures the nuclear heating in various locations and materials, and interpret
-        this as power. "damage-energy" is given by eV per source neutron.
-        Multiply by neutron source rate, and then divide by (number of atoms and
-        threshold displacement energy) to get the DPA.
-
-    Fluence
-        Measures # of neutrons streaming through.
-        "flux" is given in # per source particle, so multiply by # of source neutrons to
-        get the total fluence over the simulation.
-        Divide by area to get fluence in unit: cm^-2.
-
-    """
-    blanket_cells = [*chain.from_iterable(blanket_cell_array)]
-    div_cells = [*chain.from_iterable(divertor_cell_array)]
-    cells = blanket_cells + div_cells
-    fw_surf_cells = [
-        *(stack[0] for stack in blanket_cell_array),
-        *(stack[1] for stack in blanket_cell_array),
-    ]
-    vv_cells = [
-        *(stack[-1] for stack in blanket_cell_array),
-        *(stack[-1] for stack in divertor_cell_array),
-    ]
-    # bz_cells = [stack[2] for stack in blanket_cell_array]
-
-    # Cell filters
-    blanket_cell_filter = openmc.CellFilter(blanket_cells)
-    div_cell_filter = openmc.CellFilter(div_cells)
-    cell_filter = openmc.CellFilter(cells)
-    fw_surf_filter = openmc.CellFilter(fw_surf_cells)
-    vv_filter = openmc.CellFilter(vv_cells)
-    # bz_filter = openmc.CellFilter(bz_cells)
-
-    # material filters
-    mat_filter = openmc.MaterialFilter(material_list[:-1])
-    eurofer_filter = openmc.MaterialFilter([material_list[-1]])
-    neutron_filter = openmc.ParticleFilter(["neutron"])
-    photon_filter = openmc.ParticleFilter(["photon"])
-
-    # name, scores, filters
-    return (
-        ("TBR", "(n,Xt)", []),  # theoretical maximum TBR only, obviously.
-        # Powers
-        ("Total power", "heating", [mat_filter]),
-        ("divertor power", "heating", [div_cell_filter]),
-        ("vacuum vessel power", "heating", [vv_filter]),
-        ("breeding blanket power", "heating", [blanket_cell_filter]),
-        # Fluence
-        ("neutron flux in every cell", "flux", [cell_filter, neutron_filter]),
-        ("photon heating", "heating", [fw_surf_filter, photon_filter]),
-        # ("neutron flux in 2d mesh", "flux", [cyl_mesh_filter, neutron_filter]),
-        # TF winding pack does not exits yet, so this will have to wait
-        # DPA
-        ("eurofer damage", "damage-energy", [cell_filter, eurofer_filter]),
-        # used to get the EUROFER OBMP
-        ("divertor damage", "damage-energy", [div_cell_filter, mat_filter]),
-        ("vacuum vessel damage", "damage-energy", [vv_filter]),
-    )
-
-
-@dataclass
-class CellStage:
-    """Stage of making cells."""
-
-    blanket: list[openmc.Cell]
-    divertor: list[openmc.Cell]
-    tf_coils: list[openmc.Cell]
-    cs_coil: openmc.Cell
-    plasma: openmc.Cell
-    radiation_shield: openmc.Cell
-    ext_void: openmc.Cell
-    universe: openmc.region.Intersection
-
-    bounding_box: tuple[float, float, float, float]
-    half_bounding_box: tuple[float, float, float, float]
-
-    @property
-    def cells(self):
-        """Get the list of all cells."""
-        return (
-            *chain.from_iterable((*self.blanket, *self.divertor)),
-            *self.tf_coils,
-            self.cs_coil,
-            self.plasma,
-            self.radiation_shield,
-            self.ext_void,
-        )
-
-    @classmethod
-    def from_pre_cell_stage(
-        cls,
-        pre_cell_stage: PreCellStage,
-        tokamak_dimensions: TokamakDimensions,
-        material_library: MaterialsLibrary,
-        csg: BluemiraNeutronicsCSG,
-        tally_function: TALLY_FUNCTION_TYPE | None = None,
-        *,
-        control_id: bool = False,
-    ) -> CellStage:
-        """
-        Make CellStage from PrecellStage
-
-        Parameters
-        ----------
-        pre_cell_stage:
-            Precell Stage containing blanket and divertor pre-cells
-        tokamak_dimensions:
-            A parameter
-            :class:`bluemira.radiation_transport.neutronics.params.TokamakDimensions`,
-            Specifying the dimensions of various layers in the blanket, divertor, and
-            central solenoid.
-        control_id: bool
-            Whether to set the blanket Cells and surface IDs by force or not.
-            With this set to True, it will be easier to understand where each cell came
-            from. However, it will lead to warnings and errors if a cell/surface is
-            generated to use a cell/surface ID that has already been used respectively.
-            Keep this as False if you're running openmc simulations multiple times in one
-            session.
-
-        Returns
-        -------
-        CellStage
-        """
-        # determine universe_box
-
-        z_max, z_min, r_max, r_min = pre_cell_stage.half_bounding_box()
-
-        z_min_adj = z_min - D_TOLERANCE
-        z_max_adj = z_max + D_TOLERANCE
-        r_max_adj = r_max + D_TOLERANCE
-
-        rad_shield_wall_tk = tokamak_dimensions.rad_shield.wall
-
-        # make the universe box, incorporates the radiation shield wall
-        universe = make_universe_box(
-            csg,
-            z_min_adj - rad_shield_wall_tk,
-            z_max_adj + rad_shield_wall_tk,
-            r_max_adj + rad_shield_wall_tk,
-            control_id=control_id,
-        )
-
-        blanket = BlanketCellArray.from_pre_cell_array(
-            pre_cell_stage.blanket,
-            material_library,
-            tokamak_dimensions,
-            csg,
-            control_id=control_id,
-        )
-
-        # change the cell and surface id register before making the divertor.
-        # (ids will only count up from here.)
-        if control_id:
-            round_up_next_openmc_ids()
-
-        divertor = DivertorCellArray.from_pre_cell_array(
-            pre_cell_stage.divertor,
-            material_library,
-            tokamak_dimensions.divertor,
-            csg=csg,
-            override_start_end_surfaces=(blanket[0].ccw_surface, blanket[-1].cw_surface),
-            # ID cannot be controlled at this point.
-        )
-
-        # make the plasma cell and the exterior void.
-        if control_id:
-            round_up_next_openmc_ids()
-
-        cs, tf = make_coils(
-            csg,
-            r_min - tokamak_dimensions.cs_coil.thickness,
-            tokamak_dimensions.cs_coil.thickness,
-            z_min_adj,
-            z_max_adj,
-            material_library,
-        )
-        # make the radiation shield wall
-        # which is a hollow of the universe box
-        rad_shield = make_radiation_shield_box(
-            csg,
-            z_min_adj,
-            z_max_adj,
-            r_max_adj,
-            universe,
-            material_library,
-        )
-        plasma, ext_void = make_void_cells(
-            csg,
-            universe=universe,
-            blanket=blanket,
-            divertor=divertor,
-            central_solenoid=cs,
-            tf_coils=tf,
-            rad_shield=rad_shield,
-            control_id=control_id,
-        )
-
-        cell_stage = cls(
-            blanket=blanket.get_hollow_merged_cells(),
-            divertor=divertor.get_hollow_merged_cells(),
-            tf_coils=tf,
-            cs_coil=cs,
-            plasma=plasma,
-            radiation_shield=rad_shield,
-            ext_void=ext_void,
-            universe=universe,
-            bounding_box=pre_cell_stage.bounding_box(),
-            half_bounding_box=pre_cell_stage.half_bounding_box(),
-        )
-        cell_stage.set_volumes(blanket, divertor)
-
-        cell_stage.set_list_of_tallies(
-            blanket_cell_array=blanket,
-            divertor_cell_array=divertor,
-            material_list=[
-                getattr(material_library, f.name) for f in fields(material_library)
-            ],
-            tally_function=tally_function,
-        )
-        return cell_stage
-
-    def set_list_of_tallies(
-        self,
-        blanket_cell_array: BlanketCellArray,
-        divertor_cell_array: DivertorCellArray,
-        material_list: list[openmc.Material],
-        tally_function: TALLY_FUNCTION_TYPE | None = None,
-    ) -> list[openmc.Tally]:
-        """Set list of tallies"""
-        self.tally_func = filter_cells if tally_function is None else tally_function
-
-        tallies_list: list[openmc.Tally] = []
-        for name, scores, filters in self.tally_func(
-            material_list, blanket_cell_array, divertor_cell_array
-        ):
-            tally = openmc.Tally(name=name)
-            tally.scores = [scores] if isinstance(scores, str) else scores
-            tally.filters = filters
-            tallies_list.append(tally)
-
-        self.list_of_tallies = tallies_list
-        return tallies_list
-
-    def set_volumes(
-        self, blanket_array: BlanketCellArray, divertor_array: DivertorCellArray
-    ):
-        """
-        Sets the volume of the voids. Not necessary/ used anywhere yet.
-        """
-        ext_vertices = exterior_vertices(self.blanket, self.divertor)
-        total_universe_volume = (
-            #  top - bottom
-            (self.universe[0].surface.z0 - self.universe[1].surface.z0)
-            * np.pi
-            * self.universe[2].surface.r ** 2  # cylinder
-        )  # cm^3
-
-        # is this needed?
-        # self.universe.volume = total_universe_volume
-
-        outer_boundary_volume = to_cm3(
-            polygon_revolve_signed_volume(ext_vertices[:, ::2].T)
-        )
-        ext_void_volume = total_universe_volume - outer_boundary_volume
-        if self.tf_coils:
-            for coil in self.tf_coils:
-                ext_void_volume -= coil.volume
-        if self.cs_coil:
-            ext_void_volume -= self.cs_coil.volume
-        self.ext_void.volume = ext_void_volume
-        blanket_volumes = sum(cell.volume for cell in chain.from_iterable(blanket_array))
-        divertor_volumes = sum(
-            cell.volume for cell in chain.from_iterable(divertor_array)
-        )
-        self.plasma.volume = outer_boundary_volume - blanket_volumes - divertor_volumes
